auth 관련 Hook 수정
