+++ src/components/index.ts
@@ -1,5 +1,5 @@
 export {AuthProvider} from "./common/contextAPI/AuthProvider";
 export {AuthContext} from "./common/contextAPI/AuthContext";
-export {default as LoginForm} from "./auth/LoginForm";
+export {default as LoginForm} from "./auth/SignInForm";
 export {default as SignUpForm} from "./auth/SignUpForm";
 export {default as TodoList} from "./todo/TodoList";
